Refactor: feed의 윗부분 코드 분기 #2
