+++ components/ui/Avatar/index.tsx
@@ -4,7 +4,7 @@ import Link from "next/link";
 interface AvatarType {
   user_img: string;
   user_name: string;
-  size?: "small" | "base" | "large" | "xlarge" | "xxlarge";
+  size?: "small" | "base" | "large" | "xl" | "xxl";
 }
 
 export default function Avatar({
@@ -16,8 +16,8 @@ export default function Avatar({
     small: "w-8 h-8",
     base: "w-10 h-10",
     large: "w-12 h-12",
-    xlarge: "w-14 h-14",
-    xxlarge: "w-20 h-20",
+    xl: "w-14 h-14",
+    xxl: "w-20 h-20",
   };
 
   const avatarSize = avatarSizeVariants[size];
