+++ components/ui/Feed/FeedHeader/Title.tsx
@@ -1,0 +1,7 @@
+export default function Title() {
+  return (
+    <div className="text-[32px]/[40px]">
+      <span className="text-[#08995C]">수박이</span>와 좋은 하루 되세요!
+    </div>
+  );
+}
